+++ packages/layout/src/components/__test__/Version.snapshot.test.tsx
@@ -1,19 +1,24 @@
-import renderer from "react-test-renderer";
+import { create } from "react-test-renderer";
 import { expect, test } from "vitest";
 
 import Version from "../Version";
 
-function toJson(component: renderer.ReactTestRenderer) {
+import type {
+  ReactTestRenderer,
+  ReactTestRendererJSON,
+} from "react-test-renderer";
+
+function toJson(component: ReactTestRenderer) {
   const result = component.toJSON();
 
   expect(result).toBeDefined();
   expect(result).not.toBeInstanceOf(Array);
 
-  return result as renderer.ReactTestRendererJSON;
+  return result as ReactTestRendererJSON;
 }
 
 test("Component matches snapshot", () => {
-  const component = renderer.create(<Version version="0.0.1-local" />);
+  const component = create(<Version version="0.0.1-local" />);
   const tree = toJson(component);
   expect(tree).toMatchSnapshot();
 });