+++ packages/ui/src/components/SubPane/index.ts
@@ -1,0 +1,3 @@
+import SubPane from "./SubPane";
+
+export { SubPane };