+++ packages/ui/src/components/TabbedPanel/index.ts
@@ -1,0 +1,3 @@
+import TabbedPanel from "./TabbedPanel";
+
+export { TabbedPanel };