+++ packages/ui/vite.config.ts
@@ -23,6 +23,9 @@ export default defineConfig(({ mode }) => {
           exports: "named",
           globals: {
             react: "React",
+            "@tanstack/react-table": "reactTable",
+            "react-debounce-input": "reactDebounceInput",
+            "react-outside-click-handler": "OutsideClickHandler",
           },
         },
       },