+++ packages/user/src/supertokens/google-login.ts
@@ -9,12 +9,12 @@ const googleLogin = async (redirectUrl: string) => {
     });
 
     window.location.assign(authUrl);
-  } catch (err: any) {
-    if (err.isSuperTokensGeneralError === true) {
-      toast.error(err.message);
-    } else {
-      toast.error("Oops! Something went wrong.");
+  } catch (err) {
+    let errorMessage = "Oops! Something went wrong.";
+    if (err instanceof Error) {
+      errorMessage = err.message;
     }
+    toast.error(errorMessage);
   }
 };